Working on playback workaround
I have come across a somewhat major issue of playback not being possible for non-premium accounts, first I tried to change the way that the embed is being added twice, but I am starting to realize that without a premium account it might not be able. I am currently trying just adding a song instead of adding the playlist, and seeing if we can get the preview of that instead maybe.